requirements.txt
absl-py==0.9.0
atari-py==0.2.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.3.0
Cython==0.29.20
fasteners==0.15
future==0.18.2
glfw==1.11.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
grpcio==1.30.0
gym==0.17.2
gym-retro==0.8.0
GPUtil==1.4.0
imageio==2.8.0
importlib-metadata==1.7.0
Markdown==3.2.2
matplotlib==3.3.2
monotonic==1.5
mujoco-py==2.0.2.10
mpi4py==3.0.3
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.2.0.34
Pillow==7.1.2
protobuf==3.12.2
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
PyPrind==2.11.2
pytest==5.4.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.0
six==1.11.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
torch==1.5.1
torchvision==0.6.1
torchviz==0.0.1
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
